feat(komokana): initial commit
This is the initial commit of a minimal working version of komokana, a
daemon that I'm starting to use both as a kanata current layer
notification widget, and as a way to automatically change kanata layers
when different conditions are met in the komorebi subscriber event
stream.

I'm not sure if this will finally be merged into the master branch or
not, but for now I'll keep the code here.