fix(wm): restore full mouse resize functionality
This commit fixes a regression that was most likely introduced in #678
which changed a bunch of stuff related to window and border dimension
calculation.

See commit 4e98d7d36d392592febd90b7555caf47ceb49552 for more information
as the same approach to fix the behaviour there has been applied here.