komorebi: use a new method to focus windows
Use the same method as FancyZones to enable setting the foreground
window. This makes it possible then to remove the thread attachment
behaviors that have a number of other complex side effects, and aren't
always allowed. In addition, cleanup old focus/raise methods some, in
particular the border window is now explicitly not activated when it is
raised, as it should never be activated.