komorebi: hide the border window when unmanged windows are focused
Previously we were dropping events that don't pertain to managed
windows, with one exception in should_manage that could probably do with
further cleanup (DisplayChange). This first step fixes the latent border
window problem, where we would retain a border window when the last
managed window was closed and focus transitioned to an unmanaged window.